Change "eco" to "holiday" mode because it was misinterpreted
The Bit 1 in byte 2 off the status notification is misinterpreted as "eco" mode. But its indicate the "holiday" mode.

e.g.:

  # Status notification for day-mode (manual mode is active)
  Notification handle = 0x0421 value: 02 01 09 00 04 22 00 00 00 00 18 03 2c 24 06

  # Status notification for eco/night-mode
  Notification handle = 0x0421 value: 02 01 09 00 04 2a 00 00 00 00 18 03 2c 24 06

  # Status notification if holiday mode is enabled
  Notification handle = 0x0421 value: 02 01 0A 00 04 22 00 00 00 00 18 03 2c 24 06

  # Status notification if holiday mode ends (manual mode is now active again)
  Notification handle = 0x0421 value: 02 01 09 00 04 22 00 00 00 00 18 03 2c 24 06

You see that byte 2 is doesn't changing if you switch between day and eco/night mode. But it switch to 0x0A if you enable holiday mode and back, if holiday mode ends.

There isn't any bit with indicate the currently the "day" or the "eco"/"night" mode is active. the only possible solution for this would be to read and save the required temp, switch to both modes and read the temps, compare the temp read with required temp and at least, set again the original required temp